refactor: move dynamic styles to a separate utils file
